Automate domain name setup including ssl certificate
serverless-domain-manager does not provide a way of finding out the behind-the-scenes API Gateway domain name, which is necessary for setting up a CNAME to map from the application's domain name to the API Gateway, so I've stopped using serverless-domain-manager and instead used raw Cloudformation for setting up the API Gateway custom domain name. Also added Cloudformation to automate the process of generating SSL certificate for the custom domain name.